Вынес проверку того, что объект находится во фрустуме камеры в контроллер камеры
