--- Assets/Code/Controllers/CameraInitialization.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1f332cba4d20bb94aaed01f280402831
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
